# apps/api/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workforce/board": {
      "post": {
        "operationId": "WorkforceController_board",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceRangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkforceBoardDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
    "/api/v1/workforce/save": {
      "post": {
        "operationId": "WorkforceController_save",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceSaveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkforceRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
    "/api/v1/workforce/preview": {
      "post": {
        "operationId": "WorkforceController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceImportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkforcePreviewDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
    "/api/v1/workforce/import": {
      "post": {
        "operationId": "WorkforceController_commit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceCommitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkforceResultDto"
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
    "/api/v1/workforce/history": {
      "post": {
        "operationId": "WorkforceController_history",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkforceHistoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkforceRevisionDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Workforce"
        ]
      }
    },
...
    "/api/v1/handover/equipment": {
      "post": {
        "operationId": "HandoverController_equipment",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverEquipmentRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverEquipmentPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    },
    "/api/v1/handover/query": {
      "post": {
        "operationId": "HandoverController_query",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverSelectionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverPageDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    },
    "/api/v1/handover/entries": {
      "post": {
        "operationId": "HandoverController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverEntryDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    },
    "/api/v1/handover/change": {
      "post": {
        "operationId": "HandoverController_change",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverChangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverEntryDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    },
    "/api/v1/handover/history": {
      "post": {
        "operationId": "HandoverController_history",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverHistoryRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverHistoryDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    },
    "/api/v1/handover/remove": {
      "post": {
        "operationId": "HandoverController_remove",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoverRemoveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoverEntryDto"
                }
              }
            }
          }
        },
        "tags": [
          "Shift Handover"
        ]
      }
    }
  },
  "info": {
    "title": "IOP local API host",
    "description": "Health plus explicitly activated local analytical POC. Business operations require a configured local demo session, current site permissions and exact configured scope; no shared-user authentication claim.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "WorkforceRangeDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          }
        },
        "required": [
          "from",
          "to"
        ]
      },
      "WorkforcePersonDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "profile"
        ]
      },
      "WorkforceRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "settings",
              "worker",
              "schedule",
              "assignment"
            ]
          },
          "revision": {
            "type": "number"
          },
          "deleted": {
            "type": "boolean"
          },
          "personName": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "description": "Kind-specific Workforce payload; see docs/development/workforce.md."
          }
        },
        "required": [
          "id",
          "kind",
          "revision",
          "deleted",
          "personName",
          "data"
        ]
      },
      "WorkforceBoardDto": {
        "type": "object",
        "properties": {
          "actorId": {
            "type": "string"
          },
          "timeZone": {
            "type": "string"
          },
          "canPlan": {
            "type": "boolean"
          },
          "canAdminister": {
            "type": "boolean"
          },
          "people": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkforcePersonDto"
            }
          },
          "settings": {
            "type": "object"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkforceRecordDto"
            }
          }
        },
        "required": [
          "actorId",
          "timeZone",
          "canPlan",
          "canAdminister",
          "people",
          "settings",
          "records"
        ]
      },
      "WorkforceSaveDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "settings",
              "worker",
              "schedule",
              "assignment"
            ]
          },
          "expectedRevision": {
            "type": "number"
          },
          "deleted": {
            "type": "boolean"
          },
          "data": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "kind",
          "expectedRevision",
          "deleted",
          "data"
        ]
      },
      "WorkforceImportDto": {
        "type": "object",
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "csv",
              "email"
            ]
          },
          "text": {
            "type": "string",
            "maxLength": 80000
          },
          "userId": {
            "type": "string",
            "description": "Explicit target user for an email; CSV has userId per row."
          }
        },
        "required": [
          "format",
          "text",
          "userId"
        ]
      },
      "WorkforcePreviewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "number"
          },
          "data": {
            "type": "object"
          },
          "outcome": {
            "type": "string",
            "enum": [
              "create",
              "replace",
              "unchanged"
            ]
          }
        },
        "required": [
          "id",
          "expectedRevision",
          "data",
          "outcome"
        ]
      },
      "WorkforceImportRevisionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "expectedRevision"
        ]
      },
      "WorkforceCommitDto": {
        "type": "object",
        "properties": {
          "input": {
            "$ref": "#/components/schemas/WorkforceImportDto"
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WorkforceImportRevisionDto"
            }
          }
        },
        "required": [
          "input",
          "revisions"
        ]
      },
      "WorkforceResultDto": {
        "type": "object",
        "properties": {
          "changed": {
            "type": "number"
          },
          "unchanged": {
            "type": "number"
          }
        },
        "required": [
          "changed",
          "unchanged"
        ]
      },
      "WorkforceHistoryRequestDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "settings",
              "worker",
              "schedule",
              "assignment"
            ]
          },
          "id": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "id"
        ]
      },
      "WorkforceRevisionDto": {
        "type": "object",
        "properties": {
          "record": {
            "$ref": "#/components/schemas/WorkforceRecordDto"
          },
          "actorId": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "at": {
            "type": "string"
          },
          "action": {
            "type": "string"
          }
        },
        "required": [
          "record",
          "actorId",
          "actorName",
          "at",
          "action"
        ]
      },
...
      "HandoverEquipmentRequestDto": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "string"
          },
          "areaId": {
            "type": "string"
          },
          "search": {
            "type": "string"
          },
          "after": {
            "type": "string"
          }
        },
        "required": [
          "departmentId",
          "areaId",
          "search",
          "after"
        ]
      },
      "HandoverEquipmentPageDto": {
        "type": "object",
        "properties": {
          "codes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "nextCursor": {
            "type": "string"
          }
        },
        "required": [
          "codes",
          "nextCursor"
        ]
      },
      "HandoverSelectionDto": {
        "type": "object",
        "properties": {
          "mine": {
            "type": "boolean"
          },
          "attention": {
            "type": "boolean"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "areaId": {
            "type": "string"
          },
          "equipmentReferenceId": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "search": {
            "type": "string"
          },
          "cursor": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "",
              "none",
              "open",
              "in-progress",
              "resolved",
              "pending"
            ]
          },
          "highlights": {
            "type": "boolean"
          }
        },
        "required": [
          "from",
          "to",
          "departmentId",
          "areaId",
          "equipmentReferenceId",
          "categoryId",
          "search",
          "cursor",
          "state",
          "highlights"
        ]
      },
      "HandoverLatestUpdateDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "at": {
            "type": "string"
          }
        },
        "required": [
          "note",
          "actorName",
          "at"
        ]
      },
      "HandoverContentDto": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "departmentId": {
            "type": "string"
          },
          "areaId": {
            "type": "string"
          },
          "equipmentCode": {
            "type": "string"
          },
          "equipmentNamespace": {
            "type": "string"
          },
          "condition": {
            "type": "string",
            "enum": [
              "",
              "damaged",
              "inspection-needed",
              "blocked",
              "repaired",
              "restored"
            ]
          },
          "externalReference": {
            "type": "string"
          },
          "challenge": {
            "type": "string"
          },
          "cause": {
            "type": "string"
          },
          "measure": {
            "type": "string"
          },
          "dueDate": {
            "type": "string"
          },
          "feedbackDueDate": {
            "type": "string"
          },
          "discuss": {
            "type": "boolean"
          }
        },
        "required": [
          "date",
          "categoryId",
          "summary",
          "details",
          "departmentId",
          "areaId",
          "equipmentCode",
          "equipmentNamespace",
          "condition",
          "externalReference",
          "challenge",
          "cause",
          "measure",
          "dueDate",
          "feedbackDueDate",
          "discuss"
        ]
      },
      "HandoverEntryDto": {
        "type": "object",
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "latestUpdate": {
            "$ref": "#/components/schemas/HandoverLatestUpdateDto"
          },
          "id": {
            "type": "string"
          },
          "authorId": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "departmentLabel": {
            "type": "string"
          },
          "areaLabel": {
            "type": "string"
          },
          "categoryLabel": {
            "type": "string"
          },
          "equipmentReferenceId": {
            "type": "string"
          },
          "responsibleId": {
            "type": "string"
          },
          "responsibleName": {
            "type": "string"
          },
          "highlightedAt": {
            "type": "string"
          },
          "revision": {
            "type": "number"
          },
          "content": {
            "$ref": "#/components/schemas/HandoverContentDto"
          },
          "issueState": {
            "type": "string",
            "enum": [
              "none",
              "open",
              "in-progress",
              "resolved"
            ]
          },
          "highlighted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "authorId",
          "authorName",
          "createdAt",
          "updatedAt",
          "departmentLabel",
          "areaLabel",
          "categoryLabel",
          "equipmentReferenceId",
          "responsibleId",
          "responsibleName",
          "highlightedAt",
          "revision",
          "content",
          "issueState",
          "highlighted"
        ]
      },
      "HandoverPageDto": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HandoverEntryDto"
            }
          },
          "nextCursor": {
            "type": "string"
          },
          "total": {
            "type": "number"
          }
        },
        "required": [
          "entries",
          "nextCursor",
          "total"
        ]
      },
      "HandoverCreateDto": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "responsibleId": {
            "type": "string"
          },
          "content": {
            "$ref": "#/components/schemas/HandoverContentDto"
          },
          "issue": {
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "responsibleId",
          "content",
          "issue"
        ]
      },
      "HandoverChangeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "number"
          },
          "action": {
            "type": "string",
            "enum": [
              "correct",
              "follow-up",
              "state",
              "assign",
              "highlight"
            ]
          },
          "content": {
            "$ref": "#/components/schemas/HandoverContentDto"
          },
          "state": {
            "type": "string",
            "enum": [
              "none",
              "open",
              "in-progress",
              "resolved"
            ]
          },
          "responsibleId": {
            "type": "string"
          },
          "highlighted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "note",
          "expectedRevision",
          "action"
        ]
      },
      "HandoverHistoryRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "before": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "before"
        ]
      },
      "HandoverRevisionDto": {
        "type": "object",
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/HandoverEntryDto"
          },
          "actorId": {
            "type": "string"
          },
          "actorName": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "at": {
            "type": "string"
          }
        },
        "required": [
          "entry",
          "actorId",
          "actorName",
          "action",
          "note",
          "at"
        ]
      },
      "HandoverHistoryDto": {
        "type": "object",
        "properties": {
          "entry": {
            "$ref": "#/components/schemas/HandoverEntryDto"
          },
          "revisions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HandoverRevisionDto"
            }
          },
          "nextBefore": {
            "type": "number"
          }
        },
        "required": [
          "entry",
          "revisions",
          "nextBefore"
        ]
      },
      "HandoverRemoveDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "expectedRevision": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "expectedRevision"
        ]
      }
    }
  }
}